+++ caqtus/device/sequencer/__init__.py
@@ -1,10 +1,9 @@
 """This module implement the various components for sequencer devices.
 Sequencers are devices that output values at regular time intervals.
-
-Sequencers have a predetermined set of channels.
 """
 
 from . import channel_commands
+from . import instructions
 from ._controller import SequencerController
 from ._proxy import SequencerProxy
 from .compilation import SequencerCompiler
@@ -40,4 +39,5 @@ __all__ = [
     "SequencerProxy",
     "SequencerController",
     "channel_commands",
+    "instructions",
 ]