+++ caqtus/device/sequencer/instructions/__init__.py
@@ -8,8 +8,10 @@ from ._instructions import (
     join,
     concatenate,
 )
+from ._plot import plot_instruction
 from ._ramp import ramp, Ramp
 from ._stack import stack_instructions
+from ._to_graph import to_graph
 from ._to_time_array import convert_to_change_arrays
 from ._with_name import with_name
 
@@ -25,4 +27,6 @@ __all__ = [
     "concatenate",
     "ramp",
     "Ramp",
+    "plot_instruction",
+    "to_graph",
 ]