+++ changelog.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Documentation for sequencer instructions
+- Function `plot_instruction` to plot values takes by a sequencer instruction
+- Function `to_graph` to visualize sequencer instruction
+
 ## [4.16.0] - 2024-07-22
 
 ### Changed